svn path=/workspaces/didactylos/; revision=14345

+++ html/user/edit_forum_preferences_action.php
@@ -1,8 +1,5 @@
 <?php
 
-// commit the chanegs made in edit_forum_preferences_form.php to the database.
-
-
 require_once("../inc/forum.inc");
 require_once("../inc/image.inc"); // Avatar scaling
 
@@ -15,15 +12,34 @@ if (post_str("account_key", true) != null) {
 }
 BoincForumPrefs::lookup($user);
 
-// If the user has requested a total reset of preferences:
+if (post_str("action", true)=="reset_confirm"){
+    page_head("Confirm reset");
+    echo "This action will erase any changes you have made
+        in your community preferences.
+        To cancel, click your browser's Back button.
+        <p>
+        <form action=edit_forum_preferences_action.php method=post>
+        <input type=hidden name=action value=reset>
+        <input type=submit value=\"Reset preferences\">
+        </form>
+    ";
+    page_tail();
+    exit();
+}
+
+// If the user has requested a reset of preferences;
+// preserve a few fields.
 //
 if (post_str("action", true)=="reset"){
     $posts = $user->prefs->posts;
+    $last_post = $user->prefs->last_post;
+    $rated_posts = $user->prefs->rated_posts;
     $banished_until = $user->prefs->banished_until;
     $special_user = $user->prefs->special_user;
     $user->prefs->delete();
-    BoincForumPrefs::lookup($user);
-    $user->prefs->update("posts=$posts, banished_until=$banished_until, special_user=$special_user");
+    unset($user->prefs);
+    BoincForumPrefs::lookup($user, true);
+    $user->prefs->update("posts=$posts, last_post=$last_post, rated_posts='$rated_posts', banished_until=$banished_until, special_user='$special_user'");
     Header("Location: edit_forum_preferences_form.php");
     exit;
 }
@@ -60,7 +76,7 @@ if ($avatar_type==0){
         ImageJPEG($image2, $newfile);
     }
     if (file_exists($newfile)){
-        $avatar_url=IMAGE_URL.$user->getID()."_avatar.jpg"; //$newfile;
+        $avatar_url=IMAGE_URL.$user->id."_avatar.jpg"; //$newfile;
     } else {
         //User didn't upload a compatible file or it went lost on the server
         $avatar_url="";
@@ -76,8 +92,8 @@ $jump_to_unread = ($_POST["forum_jump_to_unread"]!="")?1:0;
 $ignore_sticky_posts = ($_POST["forum_ignore_sticky_posts"]!="")?1:0;
 $no_signature_by_default = ($_POST["signature_by_default"]!="")?0:1;
 $pm_notification = ($_POST["pm_notification"]!="")?1:0;
-$low_rating_threshold = post_int("forum_low_rating_threshold");
-$high_rating_threshold = post_int("forum_high_rating_threshold");
+//$low_rating_threshold = post_int("forum_low_rating_threshold");
+//$high_rating_threshold = post_int("forum_high_rating_threshold");
 $signature = stripslashes($_POST["signature"]);
 if (strlen($signature)>250) {
     error_page(
@@ -96,7 +112,7 @@ if ($display_wrap_postcount<5) $display_wrap_postcount=5;
 
 $signature = BoincDb::escape_string($signature);
 
-$user->prefs->update("images_as_links=$images_as_links, link_popup=$link_popup, hide_avatars=$hide_avatars, hide_signatures=$hide_signatures, jump_to_unread=$jump_to_unread, ignore_sticky_posts=$ignore_sticky_posts, no_signature_by_default=$no_signature_by_default, pm_notification=$pm_notification, avatar='$avatar_url', low_rating_threshold=$low_rating_threshold, high_rating_threshold=$high_rating_threshold, signature='$signature', forum_sorting=$forum_sort, thread_sorting=$thread_sort, minimum_wrap_postcount=$minimum_wrap_postcount, display_wrap_postcount=$display_wrap_postcount");
+$user->prefs->update("images_as_links=$images_as_links, link_popup=$link_popup, hide_avatars=$hide_avatars, hide_signatures=$hide_signatures, jump_to_unread=$jump_to_unread, ignore_sticky_posts=$ignore_sticky_posts, no_signature_by_default=$no_signature_by_default, pm_notification=$pm_notification, avatar='$avatar_url', signature='$signature', forum_sorting=$forum_sort, thread_sorting=$thread_sort, minimum_wrap_postcount=$minimum_wrap_postcount, display_wrap_postcount=$display_wrap_postcount");
 
 
 $add_user_to_filter = ($_POST["add_user_to_filter"]!="");
@@ -117,7 +133,7 @@ if ($add_user_to_filter){
 $ignored_users = get_ignored_list($user);
 for ($i=0;$i<sizeof($ignored_users);$i++){
     if ($_POST["remove".trim($ignored_users[$i])]!=""){
-        $other_user = BoincUser::lookup_id($user_to_add);
+        $other_user = BoincUser::lookup_id($ignored_users[$i]);
         if (!$other_user) {
             echo "No such user: $other_user";
         } else {
@@ -129,7 +145,8 @@ for ($i=0;$i<sizeof($ignored_users);$i++){
 
 if ($rpc == false) {
     // If we get down here everything went ok
-    // so  redirect the user to the setup page again
+    // redirect the user to the setup page again
+    //
     Header("Location: edit_forum_preferences_form.php");
 } else {
     echo "<status>\n";
